UPSTREAM:PROTOCOL:hostapd: fix overwriting MLD info during auth handling
Currently whenever we handle auth frame and if the transaction is one, we
memset the mld info and read it again with the incoming frame. However,
this should be done only when kernel is also aware of the change in the
sta info. Or else, source address will be STA's MLD MAC address and it
will be copied in link address which will be wrong.

This leads to issue where in STA entry is created with its link address as
MLD MAC address and hence even after address translation, frames are sent
out on MLD MAC address which is incorrect.

Fix this by memseting the MLD info and processing it only when the STA is
going to get re-added in the kernel so that the info is same across hostapd
and the kernel.

TAG: MLME

clo-patch-file: https://git.codelinaro.org/clo/qsdk/oss/system/feeds/wlan-hostapd/-/blob/win.wlan_host_opensource.4.0.r30/patches/r36-01-hostapd-fix-overwriting-MLD-info-during-auth-handli.patch
Gerrit: https://review-android.quicinc.com/c/oss/system/feeds/wlan-hostapd/+/5508634

Change-Id: I26e01f362f297a65daaf2b29692064b50739b777
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>
Signed-off-by: Roopni Devanathan <rdevanat@qti.qualcomm.com>