UPSTREAM:PROTOCOL:wpa_supplicant: Add radar detected event wpa_msg print
When a WDS repeater detects radar, the AP VAP is brought down as channel
change is not possible since the STA VAP would still be connected to the
root AP. The STA VAP stays connected to the Root AP on the radar affected
channel till the apsta_mode script brings it down and this is a violation.

In order to prevent this, the STA VAP should be brought down as soon as it
receives the radar-detected event. Add a wpa_msg printing the reception
of DFS_EVENT_RADAR_DETECTED so that the script listening to the events
can bring down the STA VAP when it sees the msg. Also, send the radar
event to supplicant after checking if the VAP is a STA VAP.

TAG: DFS

clo-patch-file: https://git.codelinaro.org/clo/qsdk/oss/system/feeds/wlan-hostapd/-/blob/win.wlan_host_opensource.4.0.r30/patches/r04-006-wpa_supplicant-Add-radar-detected-event-wpa_msg-prin.patch
Gerrit: https://review-android.quicinc.com/c/oss/system/feeds/wlan-hostapd/+/5019409

Change-Id: I647a7ea137e76db24d4fdd05abe1c1fea9aeb908
Signed-off-by: Amith A <quic_amitajit@quicinc.com>
Signed-off-by: Roopni Devanathan <rdevanat@qti.qualcomm.com>