QSDK: [WAR] wpa_supplicant: Add changes to reassociate if associated links is lesser than requested link
With the latest hostapd, when the partner link is going for a CAC/ACS
scanning, other link(s) will start beaconing without any RNR for
affliated links. And in the station side, while scanning if it has the
entry for the requested SSID, it will first check if the beacon has all
the requested links are beaconing and try to associated immediately if
the RNR has all links information. But, if it has less number of links,
supplicant will retry for 5 times, after which it will associated with
less number of links. This leads to station association with less number
of links, though the AP will advertise the RNR with new link after the
CAC/ACS is completed, station will not be able to add that link until ML
reconfigure is supported.

Hence, add this WAR to get the existing scan entries from the kernel for
every 5 seconds if the associated number of links is lesser than the
actual user requested configuration and try to re-associate after
deauthenticating the existing link(s) if scan entry is found with the
new link needed for completing the user configured association.

TAG: WAR

clo-patch-file: https://git.codelinaro.org/clo/qsdk/oss/system/feeds/wlan-hostapd/-/blob/win.wlan_host_opensource.4.0.r30/patches/r12-WAR-wpa_supplicant-Add-changes-to-reassociate-if.patch
Gerrit: https://review-android.quicinc.com/c/oss/system/feeds/wlan-hostapd/+/5054266

Patch-depdendency: none

Patch-work: none

MLO

Change-Id: Ib6c3843d01f352d6264cded6ca5bb3af178140f6
Signed-off-by: Manish Dharanenthiran <quic_mdharane@quicinc.com>
Signed-off-by: Hari Naraayana Desikan Kannan <quic_hnaraaya@quicinc.com>
Signed-off-by: Karthik M <<quic_karm@quicinc.com>
Signed-off-by: Roopni Devanathan <rdevanat@qti.qualcomm.com>