UPSTREAM:PROTOCOL:Free STA Associated to Partner BSS on Link Disable
Currently, when a link is disabled and a STA is associated
as a partner link to the current link, a broadcast deauth is sent,
and the STA gets disconnected. However, since the STA is connected
only as a partner link (non-assoc link), the STA reference is not
deleted and is only removed after the maximum inactivity time.

This causes an issue where the STA tries to associate after deauth
but fails because the STA reference remains.

To address this issue, when a link is disabled, check if the STA
is a partner link of the currently disabled BSS and remove the STA
reference, allowing the STA to retry association.

TAG: ML_RECONFIG

clo-patch-file: https://git.codelinaro.org/clo/qsdk/oss/system/feeds/wlan-hostapd/-/blob/win.wlan_host_opensource.4.0.r30/patches/r55-Free-STA-Associated-to-Partner-BSS-on-Link-Disable.patch
Gerrit: https://review-android.quicinc.com/c/oss/system/feeds/wlan-hostapd/+/5695768

Change-Id: I9faec601c0568ce7e7ff95d15d577ccb576d6ed9
Signed-off-by: Hari Naraayana Desikan Kannan <quic_hnaraaya@quicinc.com>
Signed-off-by: Roopni Devanathan <rdevanat@qti.qualcomm.com>