UPSTREAM: PROTOCOL: Fix legacy station assoc failure
Currently, when a station tries to associate as a multi-link (ML)
connection but fails to connect successfully
(station disassociated/deauthed and marked as not authorized),
we maintain the station context until the maximum inactivity timer
expires. If that particular station tries to re-associate as a legacy
station on a different link (non-associated link), the association
fails after EAPOL because we have not cleared the associated link
context. The existing code clears the link station context during
authentication only if the station is associated as a multi-link (ML)
connection.Also, if the station information is present on the partner
link, the station entry is removed as part of pruning after sending
disassociation on the first radio where it was originally associated.
With MLO, all links of the same MLD use the same netdev, so pruning
the legacy station removes the kernel entry completely, leading to
repeated association and pruning until the station context is removed
due to maximum inactivity on the first link the station was associated.

To address this issue, clear the partner link for any station
association, provided the partner link is not authorized.
Also skipping pruning if the station is authorized and tries to
associate in partner links of the same MLD. This change
ensures that the station can successfully associate as a legacy
station on a different link without encountering association failures.

TODO: Currently station is freed after max inactivity. But we need
to remove after disassoc timer handler. Need to cleanup and handle
scenarios like SA Query for MLD AP and legacy stas roaming.

TAG: MLME

clo-patch-file: https://git.codelinaro.org/clo/qsdk/oss/system/feeds/wlan-hostapd/-/blob/win.wlan_host_opensource.4.0.r30/patches/r58-UPSTREAM-PROTOCOL-Fix-legacy-station-assoc-failure.patch
Gerrit: https://review-android.quicinc.com/c/oss/system/feeds/wlan-hostapd/+/5772490

Change-Id: Ic58433aab20d3dcf2533abe7ae1386657d19413c
Signed-off-by: Hari Naraayana Desikan Kannan <quic_hnaraaya@quicinc.com>
Signed-off-by: Arunpandi Kannan <quic_arunpand@quicinc.com>
Signed-off-by: Roopni Devanathan <rdevanat@qti.qualcomm.com>