UPSTREAM: PROTOCOL: hostapd: Add support for 6 GHz dynamic puncturing
Currently, if any of the 20 MHz subchannels in the operating
bandwidth of a 6 GHz AP interface is marked as disabled or no IR
for 6 GHz standard power (SP) mode, the channel is considered
unavailable for SP operation. This is suboptimal as we operate
in lower power modes even though such a channel can be used
for SP operation, if the disabled or no IR subchannel is
punctured.

Also, if any subchannel in the operating bandwidth has very low
SP EIRP based on the AFC payload, the EIRP of the entire
bandwidth gets reduced due to few subchannels having lower power.
The EIRP of the channel can be enhanced if the subchannel
with lower EIRP is punctured.

Add support for dynamic puncturing of 6 GHz standard power channels
in hostapd. This allows the AP to dynamically puncture the
subchannels that meet the criteria for dynamic puncturing:
 - The subchannel which does not support SP operation due to regulatory
    restrictions.
 - The subchannel which is marked as disabled or no IR
   for 6 GHz SP mode.
 - The subchannel which has very low EIRP based on the AFC
   payload.

The resulting transmit power after further puncturing must exceed both
    (1) the unpunctured channel EIRP for SP mode and
    (2) the unpunctured channel EIRP of LPI or VLP modes (if best power
        mode is enabled).
If best power mode is disabled, only condition (1) is checked.

If the user has provided a puncture pattern, dynamic puncturing will
try to find a puncture pattern that also intersects with the
user-provided puncture pattern.

Use a balanced puncturing approach such that the puncturing
is done in a way that minimizes the impact on the overall bandwidth
while still allowing for higher power operation.

Dynamic puncturing is enabled by default and can be turned off
using the iface config 'puncture_strict_6ghz' option.
When this option is set to 1, the AP will not puncture any subchannels.
User given puncture pattern will be used as is.

TAG: AFC

clo-patch-file: https://git.codelinaro.org/clo/qsdk/oss/system/feeds/wlan-hostapd/-/blob/win.wlan_host_opensource.4.0.r30/patches/r82-a-UPSTREAM-PROTOCOL-hostapd-Add-support-for-6-GHz-dyna.patch
Gerrit: https://review-android.quicinc.com/c/oss/system/feeds/wlan-hostapd/+/6106709

Fixes: None
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Patch-dependency:
n00-001-hostapd-add-support-for-6GHz-operation.patch
r32-001-hostapd-Add-support-to-receive-AFC-power-update-comp.patch
r72-a-UPSTREAM-PROTOCOL-hostapd-6-GHz-power-mode-channels-.patch
r72-d-UPSTREAM-PROTOCOL-hostapd-Clean-up-and-refactor-6-GH.patch
r72-e-UPSTREAM-PROTOCOL-hostapd-Support-for-SP-and-Punctur.patch
r72-h-UPSTREAM-PROTOCOL-hostapd-Apply-best-power-mode-for-.patch
r80-d-UPSTREAM-PROTOCOL-hostapd-Add-minimum-PSD-calculatio.patch
r81-UPSTREAM-PROTOCOL-hostapd-Use-OOBE-compliant-PSD-for.patch

Change-Id: I45d82e0d51b1fab1d38c15c605d7c809bad02b0a
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>
Signed-off-by: Roopni Devanathan <rdevanat@qti.qualcomm.com>