wpa_supplicant: Fix missing update of 6G assoc freq
When the associated frequency is not set (i.e., 0) in the assoc_info structure,
`wpa_cli status` incorrectly reports a frequency of 0 for STA VAPs on 6GHz radios.
This occurs despite the BSSID being known and the frequency being available in
the BSS table.

This patch addresses the issue by checking if the frequency is zero and the BSSID
is known. If so, it retrieves the latest BSS entry for the BSSID and updates the
assoc_info.freq accordingly. This ensures that the correct frequency is reported
in the status output.

TAG: GENERAL

clo-patch-file: https://git.codelinaro.org/clo/qsdk/oss/system/feeds/wlan-hostapd/-/blob/win.wlan_host_opensource.4.0.r30/patches/r96-wpa_supplicant-Fix-missing-update-of-6G-assoc-freq.patch
Gerrit: https://review-android.quicinc.com/c/oss/system/feeds/wlan-hostapd/+/6129948

Change-Id: Ic51276d85a8cc05368b738c46cbb9e9e53158155
Signed-off-by: Banka Sai Krishna <banksaik@qti.qualcomm.com>
Signed-off-by: Roopni Devanathan <rdevanat@qti.qualcomm.com>