UPSTREAM:PROTOCOL:hostapd:Fix multiple ANonce generation during repeated FT auth
During roaming, if the AP delays its authentication response,
the station may retry the authentication request. The AP
responds to each retry, generating a new ANonce for every
authentication reply. However, the station uses the ANonce
from the first response to construct the FTIE in its
reassociation request. This mismatch causes the AP to reject
the reassociation due to an invalid FTIE.

This patch fixes the issue by ensuring that the AP generates
the ANonce only once per FT authentication attempt and
reuses the same ANonce for all authentication replies to the
same station. This ensures consistency in the FT handshake
and prevents reassociation failures.

Change-Id: I00bd0e139be81b6a7e1280f6ec915f4831fe8b26
Signed-off-by: Shiva Sankar Gajula <sgajula@qti.qualcomm.com>
Signed-off-by: Sowmiya Sree Elavalagan <ssreeela@qti.qualcomm.com>
Signed-off-by: Roopni Devanathan <rdevanat@qti.qualcomm.com>