UPSTREAM:PROTOCOL:hostapd: Run group rekey on the first authenticator to enter SETKEYS
When the GTK rekey timer fires on the primary authenticator, new group
keys are generated for all links in an AP MLD and new keys are updated
to the associated stations via a group key handshake. This is achieved
by moving the group state for each authenticator in an AP MLD to the
SETKEYS state one after another. In the multi link case, new group keys
for all setup links are updated to stations in one shot via a single
group handshake message 1 sent on the association link. So group key
update to stations is skipped on non-association links and those cases
simply return from wpa_group_update_sta(). At this point
GKeyDoneStations is zero and group->changed is true on the
non-association link. Due to the group->changed value being set, the
do-while loop in wpa_rekey_all_groups() causes the group state machine
for the non-association link to jump to the next state, SETKEYSDONE, and
install the new keys to the driver/firmware on the non-association link
even before they could be communicated to the stations with group
handshake message 1. This would mean that the partner link Beacon frames
could be sent using a new BIGTK that is not yet known by the associated
stations posing the risk for a potential beacon miss due to beacon
protection validation failures.

Prevent the partner authenticator from jumping to the SETKEYSDONE state
before the new group keys are sent to the associated stations, run rekey
on the first authenticator to enter SETKEYS state, i.e, instead of
returning from wpa_group_update_sta() for non-association links, let the
first authenticator entering SETKEYS initiate transition to
REKEYNEGOTIATING to update keys to stations regardless of whether it is
the association link. This means that when the first authenticator in
the AP MLD enters SETKEYS state from wpa_rekey_gtk(), GKeyDoneStations
count is incremented in the group state machine data of each of the
setup links and GUpdateStationKeys is set and a call to wpa_sm_step() is
made so the wpa_ptk_group state machine can transition to the
REKEYNEGOTIATING state and the group handshake messages can be exchanged
with the station to communicate the group keys of all the setup links in
group handshake message 1. The next authenticator in the AP MLD to enter
wpa_group_update_sta() will simply return seeing sm->wpa_ptk_group_state
= WPA_PTK_GROUP_REKEYNEGOTIATING. The non-zero GKeyDoneStations value in
each of the setup links will in turn ensure that no authenticator jumps
to the SETKEYSDONE state before rekey is established with all the
associated stations.

TAG: GTK_REKEY
Patch-dependency: None

Change-Id: I69881c2cacbd9c2802b0fc83c2cba3bb5cf4feba
Co-developed-by: Adil Saeed Musthafa <adilm@qti.qualcomm.com>
Signed-off-by: Adil Saeed Musthafa <adilm@qti.qualcomm.com>
Signed-off-by: Sai Pratyusha Magam <smagam@qti.qualcomm.com>