UPSTREAM:PROTOCOL:hostapd:nl80211: Support scanning as AP when not beaconing
If the kernel indicated support for scanning on AP interface, set
NL80211_SCAN_FLAG_AP in the scan request when the interface is not yet
beaconing. In that case, it is possible to do a scan without interfering
with normal AP operation so it is safe to set the flag.

This is important because otherwise it may not be possible to do the
appropriate scans required to bring up an interface that is part of an
MLD.

TAG:DOWNSTREAM

Feature: Scan

Patch-work: https://git.w1.fi/cgit/hostap/commit/?id=ba9dacb42c4de8f59a6c7a8a5d3e98b2cc535e8d
Patch-dependency: None

Change-Id: I238b641338adb06cb5414f982b9b49f8dc1cec8e
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-by: Andrei Otcheretianski <andrei.otcheretianski@intel.com>