UPSTREAM:PROTOCOL:hostapd:wpa_debug: Use separate buffer for path and improve error checking
Using the same buffer for output and input could already result in
an overlapping source and destination in snprintf. This was working
fine, however, with the patch to also find tracefs, we continue to parse
the buffer. In that case, the continued parsing can corrupt the found
path causing an error.

Fix this problem and reshuffle the code a bit to make it a bit more
clear and improve the condition to skip lines that could not be parsed
properly.

TAG: DOWNSTREAM

Feature: DEBUG

Patch-work: https://git.w1.fi/cgit/hostap/commit/?id=672da5df77220684b9bbf3de5f5411c42a1e6380
Patch-dependency: None

Fixes: 0a76c7ed64de ("wpa_debug: Prefer tracefs over debugfs")
Change-Id: I8e63e64432724de9e7cb2ec1f6ca4bc5946848f9
Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>