UPSTREAM:PROTOCOL:hostapd:utils: Keep the last NOTE message as context for backtraces
When storing a backtrace, keep the string that was last used in a NOTE
message as context information. Keep all past strings in a list so that
they will remain available.

With this, we are able to show in leak reports which test did the
allocation. This is useful to more quickly solve issues as it can be
hard to find the triggering test otherwise.

TAG: DOWNSTREAM

Feature: DEBUG

Patch-work: https://git.w1.fi/cgit/hostap/commit/?id=24c6b9e4f672c74fecb7c1b006764d766d5c05a8
Patch-dependency: None

Change-Id: Ib46bd59d800130ccd3108fbe8e057ef5a3ce033d
Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>