UPSTREAM:PROTOCOL:wpa_supplicant:P2P2: Handle identity ID consistently within files
Using an ID to reference a specific identity= block within the same file
for the DIK has the problem that it requires the IDs to be stable.
However, we also use a static int for the ID enumeration in case
multiple files are read.

Overall, it seems like it could be better to use the DIK as an
identifier. However, we do not, so the numbering needs to be restored
consistently within the file. The straight forward way to achieve this
is to store the base of the counter as a static variable and adding it
to both the go_dik_id and the id of the identity= blocks.

For all of this to work, we also need to make sure that we write out the
values with file-local indices. As such, we need to use the position in
the internal list rather than the assigned ID when writing the values.

TAG: DOWNSTREAM

Feature: P2P

Patch-dependency: None
Patch-work: https://git.w1.fi/cgit/hostap/commit/?id=cad3af08e2aeea88b69cb25d40ea659f85a42106

Fixes: 417c67468b8d ("P2P2: Add device identity block to wpa_supplicant configuration")
Fixes: ec4569174750 ("P2P2: Store ID of Device Identity block in network block")
Change-Id: I1b63f2443a1d8d07dd304699a8dc95311e673673
CC: Shivani Baranwal <quic_shivbara@quicinc.com>
CC: Vinay Gannevaram <quic_vganneva@quicinc.com>
Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>