QSDK:PROTOCOL: hostapd: Channel switch to same channel for DFS test mode
When a 5 GHz link is configured in DFS test mode for probability
detection, the expected behavior is to send a Channel Switch
Announcement (CSA) where the next channel in the CSA IE is the same
as the current channel.

Currently, hostapd achieves this by updating the beacon template with
a fixed CSA IE and a hardcoded CSA count of 3. After 1 second, hostapd
removes the CSA IE and sends a new beacon template to the driver.

However, because the CSA IE is added to a static template rather than
through an actual CSA operation, the CSA offsets are not populated when
the beacon template is sent to the firmware. With recent firmware
validation checks, this results in an error: the template contains
CSA IEs, but the CSA offsets are incorrect. Consequently, the firmware
suspends transmissions, causing DFS probability detection tests to fail.

To fix this, instead of injecting a fixed beacon template with CSA IE,
initiate a channel switch to the same channel. This ensures CSA offsets
are computed correctly, as they would be during a normal channel switch.

For channel switch, use a CSA count of 5 which is used for other
hostapd initiated channel switches instead of current value of 3
for consistency.

TAG: DFS

Gerrit: https://review-android.quicinc.com/c/oss/system/feeds/wlan-hostapd/+/6411442

Patch-dependency:
a00-004-Add-provision-to-test-RADAR-detection-probablity-in-.patch
Fixes: I3d68b12ec8b ("hostapd: Add provision to test RADAR detection probablity in dfs testing")

Change-Id: Id6e3877b65f0c6794ca3687b007ee1594a3a9bb5
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>
Signed-off-by: Roopni Devanathan <rdevanat@qti.qualcomm.com>