UPSTREAM:PROTOCOL:hostapd: Fix SAE Auth Failure When Link Addresses Swap in Multi-Link STA
When a station initially connects as a 3-link STA (e.g., link 0 on
2.4 GHz, link 1 on 5 GHz, and link 2 on 6 GHz), it may later move
away and return without disconnecting.

In some cases, MLD stations swap link addresses and use the MAC
address of link 0 in the authentication frame for the 5 GHz band.
In the current design, during SAE confirmation, the process attempts
to fetch the MLD MAC address by looking up the station using the link
address.

However, when the link 0 address is not part of the 5 GHz hapd, the
lookup fails, causing SAE confirmation to fail. This results in an
incorrect "non-ML auth" error and adds the STA as a new non-ML
STA entry.

Fetch the MLD address directly from the authentication frame in
the SAE confirm fixed-length field instead of relying on link
address lookup.

TAG: MLME

Gerrit: https://review-android.quicinc.com/c/oss/system/feeds/wlan-hostapd/+/6415039

Patch-dependency: None

Change-Id: Ief07565ed3cc742d505423a377f4e8d2314fb051
Signed-off-by: Ramasamy Kaliappan <rkaliapp@qti.qualcomm.com>
Signed-off-by: Roopni Devanathan <rdevanat@qti.qualcomm.com>