UPSTREAM:PROTOCOL:hostapd: Use appropriate VLAN group instead of default VLAN group
The current implementation in the GTK Rekey path refers to
the default VLAN group instead of the VLAN to which the
client is bound to. For instance, the GKeyDoneStations
count is incorrectly incremented/decremented for partner
links in the default VLAN group always, even when the
station has associated on the non-zero VLAN. This can
result in inconsistencies during GTK Rekey for stations
connected on non-zero VLAN where the new group keys are
installed to the driver even before they could be
communicated to the stations on these VLANs through a
group handshake message 1.

At all of those places, let the correct VLAN group be
referenced based on the vlan_id to which the station is
connected to, instead of using the default VLAN.

Patch-dependency: Ief2639b4c0a3005e1035c87d23c591931b98691f ("UPSTREAM:PROTOCOL:hostapd: VLAN group state machine data creation for partner AP MLD links")

Change-Id: I516fe072ae210472cb481c7eb1ed6cb041eaea54
Signed-off-by: Sai Pratyusha Magam <smagam@qti.qualcomm.com>