UPSTREAM:PROTOCOL:hostapd: Re-Initialize group keys if it is the first sta seen in the entire MBSS group
In the current implementation, group keys are Re-initialized
for a first station association (first_sta_seen) in the
BSS. In a scenario where the non-transmitting BSS sees a client
association before the transmitting BSS, group keys(GTK/IGTK)
are re-initialized in wpa_gtk_update() for the non-Tx BSS.
It is checked if a BIGTK key is already set on the Tx BSS
(bigtk_set). Since BIGTK was set for the Tx BSS during its
bringup time, it simply returns from wpa_gtk_update()
for the non-Tx BSS.

If this is followed by a first station association on the Tx
BSS, a new BIGTK key is generated for the Tx BSS in wpa_gtk_update()
and the same is installed to the driver. This would mean that the
Beacon frames are being sent using the new BIGTK that is not known
to the stations associated to the non-Tx BSSes, posing the risk for
a potential beacon miss due to beacon protection validation failures.

Fix this by allowing re-initialization of group keys if and only if
it is a first station association seen across the entire MBSS group.

Patch-dependency: None
TAG: MBSSID

Change-Id: I0d2fd31447c1690950f6c8ec3e91c591ec72a65b
Signed-off-by: Sai Pratyusha Magam <smagam@qti.qualcomm.com>