UPSTREAM:PROTOCOL:hostapd: update beacons for TX profile
When nontx profiles are brought up, following warning is shown
on the console:
"ath12k_wifi7_pci 0001:01:00.0: nontransmitted profile not found in
beacon template"
This happens because when hostapd sends NL80211_CMD_START_AP to
kernel for the new profile, mac80211 still uses the old Beacon
template saved for the tx profile of the group which does not
include the new profile.

Update the Beacon template for the tx interface in
__ieee802_11_set_beacon() when a new nontransmitted profile
is getting brought up.

TAG: MBSSID

Patch-dependency: none

Change-Id: I0cf89289b7d50ef10894046237647dc187064764
Signed-off-by: Aloka Dixit <aloka.dixit@oss.qualcomm.com>