UPSTREAM:PROTOCOL:hostapd: remove stopped non-tx profile from beacons
Hostapd_cli command 'stop_ap' brings down specified interface
but if this interface is a non-transmitted (nontx) profile in
a multi-BSSID group then it is does not get removed from the
Beacon frames. This happens because only one Beacon template
is used for all profiles from the MBSSID group and it is stored
for the transmitted profile (tx) which is currently not updated.

Fix this issue by changing following:
(1) Update the Beacon template for the tx profile when a nontx
profile is stopped.
(2) Above alone is not enough because template generation code
in hostapd_eid_mbssid_elem_len() and hostapd_eid_mbssid_elem()
check only for 'started' flag while adding nontxt profiles.
This flag is not reset to 0 with 'stop_ap'. Instead check flag
'beacon_set_done' in to identify a stopped BSS.

TAG: MBSSID

Patch-dependency: none

Change-Id: I744e75d6456b6b8985acc08affdf26c9379f2ba4
Signed-off-by: Aloka Dixit <aloka.dixit@oss.qualcomm.com>