UPSTREAM: HOSTAPD: Recitfy current operating class field
The current operating class field in the supported operating class IE
present in the association request sent by a station to an AP operating
on 80 MHz on 6 GHz band (133 is the opclass for 80 MHz on 6 GHz band)
is always 131 (opclass for 20 MHz on 6 GHz band).

On the STA side, the wpas_supp_op_class_ie fills the current operating
class field in the supported operating class IE based on the HT/VHT
capabilities advertised in the beacon by the Root AP. Currently, the
HE/EHT capabilities/operation IE advertised by the Root AP are not parsed,
resulting in the 6 GHz operating class for 20 MHz BW (opclass 131) being
derived as the "current operating class".

To address this issue, parse the HE/EHT operation IE of the Root AP,
determine the Root AP's bandwidth, and then derive the
current operating class.

Tag: 6G_REGULATORY
Patch-dependency: None
Link: [PATCH] Fill current operating class field in supported operating class IE

Change-Id: I160820e68c565b1faa4a2f8020b5fbf45370f734
Signed-off-by: Hariharan Basuthkar <hbasuthk@qti.qualcomm.com>