UPSTREAM:PROTOCOL:hostapd: SME: Do not send RSNXE for WPA 1 connections
The element is not applicable to WPA 1 and it will not be included in
the KDE later on. However, hostapd checks that the KDE and the
association request match and deauthenticates the station if it sent the
RSNXE in the association request.

So omit sending it if the protocol is WPA_PROTO_WPA.

TAG: DOWNSTREAM

Feature: None

Patch-work: https://git.w1.fi/cgit/hostap/commit/?id=f531b6d2092aebd7943801e26e9e1fd00f7f826b

Patch-dependency: None

Change-Id: Idae653360fdf79ce4e18d9adf970907d687b5ff2
Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>
Signed-off-by: Karthik M <karm@qti.qualcomm.com>