QSDK:PROTOCOL: hostapd: Fix 320 MHz BW DFS test mode channel switch
When a 5 GHz link is configured in DFS test mode for probability
detection, the expected behavior is to do a Channel Switch with the
next channel same as that of the current channel.

This works fine for all the legacy bandwidths. But for 320 MHz BW,
which is not as per standard, the CSA center frequency passed to
kernel as part of the channel switch command is that of the 160 MHz
channel (VHT BW center frequency). But the width is passed correctly
as 320 MHz. But to this frequency / width mismatch, kernel rejects
the channel change and the channel switch is a failure.

To fix this, pass the current operating BW center frequency
instead of the VHT bandwidth center frequency.

Tag: DFS
Fixes: Id6e3877b65f ("QSDK:PROTOCOL: hostapd: Channel switch to same channel for DFS test mode")
Patch-dependency:
Id6e3877b65f ("QSDK:PROTOCOL: hostapd: Channel switch to same channel for DFS test mode")
I3d68b12ec8b ("hostapd: Add provision to test RADAR detection probablity in dfs testing")

Change-Id: I6e83243dbda588d5e08ecf22801896f2b0a21535
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>