QSDK:PROTOCOL:HOSTAPD:extn: Add config option for 6 GHz ESS advertisement with 2.4/5 GHz colocated APs
Introduce a configuration option (rnr_member_ess_colocated_en) to control the
"Member of ESS with 2.4/5 GHz colocated AP" bit in the BSS Parameters field of
the TBTT Info element within the RNR IE for 6 GHz radios.

By default, this bit is set to 0. When enabled, and if colocated 2.4/5 GHz BSS
instances are active, the bit indicates that the 6 GHz AP is part of an ESS that
includes colocated APs in other bands. This setting is not allowed for 6 GHz-only
APs, as they do not have colocated APs.

Configuration details:
- `rnr_member_ess_colocated_en`: Valid values are 0 or 1.
- Can be set directly in `hostapd.conf` under the radio configuration.
- Can be controlled dynamically via CLI:
   hostapd_cli -i <interface> SET rnr_member_ess_colocated_en 1
- Current value can be verified using:
    hostapd_cli -i <interface> STATUS
    rnr_member_ess_colocated_en=1

TAG: MBSSID

Patch-dependency: Ib74437d5455 (QSDK:PROTOCOL:HOSTAPD: Control RNR
addition based on colocated configuration)

Change-Id: I1fa1196a2a25ff5419c119bd9e537104d5221a2c
Signed-off-by: Basamma Yakkanahalli <ybasamma@qti.qualcomm.com>