UPSTREAM:PROTOCOL:wifi: mld: fix invalid memory access during 6 GHz reauth without deauth
When a multi-link station joins on 5 GHz + 6 GHz and then starts
authentication on 6 GHz as a legacy STA without sending a deauth,
an invalid memory access occurs during re-add.

This happens because mld_info is zeroed before the station state
machine is cleared and the reference STA is set to NULL. When mld_info is set to zero before then the clear_wpa_sm_for_each_partner_link() will not set NULL to partner links. As a result, stale pointers can be dereferenced during re-add.

Fix this by moving the memset of mld_info after clearing the old
state machine and nullifying the reference STA.

Patch-dependency: None
TAG: SECURITY

Change-Id: Icc65062063bc8c97b0355d99c47e81dfa4dd9d29
Signed-off-by: Harshitha Prem <hprem@qti.qualcomm.com>